Apple Sucks

+++ create_exe.py
@@ -29,29 +29,31 @@ def make_exe(python_file: str):
         if os.path.isdir(cur_dir + '__pycache__/'):
             shutil.rmtree(cur_dir + '__pycache__/')
 
-    # Remove old builds if any
-    clean_up()
-    # Build Exe
-    PyInstaller.__main__.run([python_file, '--onefile', '-w'])
-    # Move and Rename
-    if os.path.isfile(cur_dir + 'dist/' + python_file_name + '.exe'):
-        if not os.path.isfile(cur_dir + 'C4IconSwapper.' + version + '.exe'):
-            shutil.copy(cur_dir + 'dist/' + python_file_name + '.exe', cur_dir + 'C4IconSwapper.' + version + '.exe')
-        else:
-            os.remove(cur_dir + 'C4IconSwapper.' + version + '.exe')
-            shutil.copy(cur_dir + 'dist/' + python_file_name + '.exe', cur_dir + 'C4IconSwapper.' + version + '.exe')
-    # Remove build files
-    clean_up()
+    if not on_mac:
+        # Remove old builds if any
+        clean_up()
+        # Build Exe
+        PyInstaller.__main__.run([python_file, '--onefile', '-w'])
+        # Move and Rename
+        if os.path.isfile(cur_dir + 'dist/' + python_file_name + '.exe'):
+            if not os.path.isfile(cur_dir + 'C4IconSwapper.' + version + '.exe'):
+                shutil.copy(cur_dir + 'dist/' + python_file_name + '.exe', cur_dir + 'C4IconSwapper.' + version + '.exe')
+            else:
+                os.remove(cur_dir + 'C4IconSwapper.' + version + '.exe')
+                shutil.copy(cur_dir + 'dist/' + python_file_name + '.exe', cur_dir + 'C4IconSwapper.' + version + '.exe')
+        # Remove build files
+        clean_up()
 
 
 # Change to True before running; this is to prevent accidental run; add tkdnd2.8\ to Python\Python38\tcl\tcl8.6\
 execute = False
 
 if execute:
-    overwrite_file = '_'
-    if os.path.isfile(cur_dir + 'C4IconSwapper.' + version + '.exe'):
-        overwrite_file = input('Overwrite file? (y/n)... ')
-        if overwrite_file[0] == 'y' or overwrite_file[0] == 'Y':
+    if not on_mac:
+        overwrite_file = '_'
+        if os.path.isfile(cur_dir + 'C4IconSwapper.' + version + '.exe'):
+            overwrite_file = input('Overwrite file? (y/n)... ')
+            if overwrite_file[0] == 'y' or overwrite_file[0] == 'Y':
+                make_exe('main.py')
+        else:
             make_exe('main.py')
-    else:
-        make_exe('main.py')
